md->pdf and Reorganize code
