+++ examples/oddecho/problem_statement/problem.sv.md
@@ -1,10 +1,6 @@
-**ECHO! Echo! Ech...**
+**EKO! Eko! Ek...**
 
 ![](echo_cave.jpg)
-<!--
-Alternatively, if you want more control:
-<img src="echo_cave.jpg" style="width=40%">
--->
 
 Du älskar att skrika i grottor för att höra dina ord ekade tillbaka till dig. Tyvärr, som en hårt arbetande mjukvaruingenjör, har du
 inte tid för att komma ut och skrika i grottor så ofta. Istället skulle du vilja implementera ett program som fungerar som en ersättning för en grotta.
