+++ problemtools/md2html.py
@@ -8,8 +8,7 @@ import re
 import json
 from typing import Optional
 
-from . import verifyproblem
-from . import problem2html
+from . import statement_common
 
 
 FOOTNOTES_STRING = '<section class="footnotes" role="doc-endnotes">'
@@ -24,7 +23,7 @@ def convert(problem: str, options: argparse.Namespace) -> None:
     problembase = os.path.splitext(os.path.basename(problem))[0]
     destfile = string.Template(options.destfile).safe_substitute(problem=problembase)
 
-    statement_path = problem2html.find_statement(problem, extension="md", language=options.language)
+    statement_path = statement_common.find_statement(problem, extension="md", language=options.language)
 
     if statement_path is None:
         raise Exception('No markdown statement found')
@@ -47,7 +46,7 @@ def convert(problem: str, options: argparse.Namespace) -> None:
     if templatepath is None:
         raise Exception('Could not find directory with markdown templates')
 
-    problem_name = _get_problem_name(problem)
+    problem_name = statement_common.get_problem_name(problem, options.language)
 
     html_template = _substitute_template(templatepath, "default-layout.html",
            statement_html=statement_html,
@@ -55,7 +54,7 @@ def convert(problem: str, options: argparse.Namespace) -> None:
            title=problem_name or "Missing problem name",
            problemid=problembase)
 
-    samples = _samples_to_html(problem)
+    samples = statement_common.samples_to_html(problem)
 
     html_template = inject_samples(html_template, samples)
     html_template = replace_hr_in_footnotes(html_template)
@@ -88,6 +87,7 @@ def handle_image(src: str) -> None:
 
 
 def json_dfs(data, callback) -> None:
+    """Traverse all items in a JSON tree, find all images, and call callback for each one"""
     if isinstance(data, dict):
         for key, value in data.items():
             # Markdown-style images
@@ -96,13 +96,6 @@ def json_dfs(data, callback) -> None:
             else:
                 json_dfs(value, callback)
 
-            # HTML-style images
-            if key == "t" and value == "RawInline":
-                image_string = data["c"][1]
-                src = re.search(r'src=["\'](.*?)["\']', image_string)
-                if src:
-                    callback(src.group(1))
-    
     elif isinstance(data, list):
         for item in data:
             json_dfs(item, callback)
@@ -141,89 +134,4 @@ def _substitute_template(templatepath: str, templatefile: str, **params) -> str:
     with open(os.path.join(templatepath, templatefile), "r", encoding="utf-8") as template_file:
         html_template = template_file.read() % params
     return html_template
-
-
-def _get_problem_name(problem: str, language: str = "en") -> Optional[str]:
-    """Load problem.yaml to get problem name"""
-    with verifyproblem.Problem(problem) as prob:
-        config = verifyproblem.ProblemConfig(prob)
-    if not config.check(None):
-        print("Please add problem name to problem.yaml when using markdown")
-        return None
-    names = config.get("name")
-    # If there is only one language, per the spec that is the one we want
-    if len(names) == 1:
-        return next(iter(names.values()))
-
-    if language not in names:
-        raise Exception(f"No problem name defined for language {language}")
-    return names[language]
-
-
-def _samples_to_html(problem: str) -> str:
-    """Read all samples from the problem directory and convert them to HTML"""
-    samples_html = ""
-    sample_path = os.path.join(problem, "data", "sample")
-    interactive_samples = []
-    samples = []
-    casenum = 1
-    for sample in sorted(os.listdir(sample_path)):
-        if sample.endswith(".interaction"):
-            lines = [f"""<table class="sample" summary="sample data">
-                       <tr>
-                          <th style="text-align:left; width:33%;">Read</th>
-                          <th style="text-align:center; width:33%;">Sample Interaction {casenum}</th>
-                          <th style="text-align:right; width:33%;">Write</th>
-                       </tr>
-                    </table>"""]
-            with open(os.path.join(sample_path, sample), "r", encoding="utf-8") as infile:
-                sample_interaction = infile.readlines()
-            for interaction in sample_interaction:
-                data = interaction[1:]
-                line_type = ""
-                if interaction[0] == '>':
-                    line_type = "sampleinteractionwrite"
-                elif interaction[0] == '<':
-                    line_type = "sampleinteractionread"
-                else:
-                    print(f"Warning: Interaction had unknown prefix {interaction[0]}")
-                lines.append(f"""<div class="{line_type}"><pre>{data}</pre></div>""")
-
-            interactive_samples.append(''.join(lines))
-            casenum += 1
-            continue
-        if not sample.endswith(".in"):
-            continue
-        sample_name = sample[:-3]
-        outpath = os.path.join(sample_path, sample_name + ".ans")
-        if not os.path.isfile(outpath):
-            continue
-        with open(os.path.join(sample_path, sample), "r", encoding="utf-8") as infile:
-            sample_input = infile.read()
-        with open(outpath, "r", encoding="utf-8") as outfile:
-            sample_output = outfile.read()
-
-        samples.append("""
-            <tr>
-                <th>Sample Input %(case)d</th>
-                <th>Sample Output %(case)d</th>
-            </tr>
-            <tr>
-            <td><pre>%(input)s</pre></td>
-            <td><pre>%(output)s</pre></td>
-            </tr>"""
-            % ({"case": casenum, "input": html.escape(sample_input), "output": html.escape(sample_output)}))
-        casenum += 1
-
-    if interactive_samples:
-        samples_html += ''.join(interactive_samples)
-    if samples:
-        samples_html += f"""
-        <table class="sample" summary="sample data">
-          <tbody>
-          {''.join(samples)}
-          </tbody>
-        </table>
-        """
-    return samples_html
